workspace/pyproject.toml
[project]
name = "tds"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "pandas>=2.3.3",
    "pillow>=11.3.0",
    "statistics>=1.0.3.5",
]
